feat(qwen3-a8w8): align serving with optimized kernels
Use BF16 KV caches and remove the obsolete per-row KV scale buffers to match the updated pypto-lib kernel ABI.

Keep Gate and Up weights in INT8 form, propagate their scales through loading, compilation, and L3 dispatch, and update placeholder tensors accordingly.

Refresh batching assertions for the split prefill embedding paths.

Validated with: pytest -q tests/test_batching.py (37 tests completed; the local NPU runtime hung during process teardown).

+++ examples/model/qwen3_14b/npu_generate.py
@@ -516,7 +516,7 @@ def main() -> None:
                 max_seq_len=args.max_seq_len,
                 max_new_tokens=args.max_new_tokens,
                 device="cpu",
-                kv_dtype="int8" if args.model_format == _QWEN3_A8W8_FORMAT else args.kv_cache_dtype,
+                kv_dtype="bfloat16" if args.model_format == _QWEN3_A8W8_FORMAT else args.kv_cache_dtype,
                 weight_dtype="bfloat16" if args.model_format == _QWEN3_A8W8_FORMAT else args.dtype,
                 npu_memory_utilization=args.npu_memory_utilization,
                 max_num_batched_tokens=args.max_num_batched_tokens,
